refactor: move encrypt_and_send_raw_to_device to encryption mod
